Fix import in generate-icons that broke everything
