+++ app/src/ui/octicons/octicons.generated.ts
@@ -1,9 +1,22 @@
 /*
- * This file is automatically generated by the generate-octicons tool.
- * Manually changing this file will only lead to sadness.
- */
+* This file is automatically generated by the generate-octicons tool.
+* Manually changing this file will only lead to sadness.
+*/
 
-import { OcticonSymbolVariants } from '.'
+export type OcticonSymbolVariant = {
+  /** SVG path element data */
+  readonly p: string[]
+
+  /** The width of the symbol */
+  readonly w: number
+
+  /** The height of the symbol */
+  readonly h: number
+}
+
+export type OcticonSymbolVariants = Record<PropertyKey, OcticonSymbolVariant>
+
+export type OcticonSymbol = OcticonSymbolVariant | OcticonSymbolVariants
 
 export const accessibility: OcticonSymbolVariants = {"16":{"p":["M9.923 5.302c.063.063.122.129.178.198H14A.75.75 0 0 1 14 7h-3.3l.578 5.163.362 2.997a.75.75 0 0 1-1.49.18L9.868 13H6.132l-.282 2.34a.75.75 0 0 1-1.49-.18l.362-2.997L5.3 7H2a.75.75 0 0 1 0-1.5h3.9a2.54 2.54 0 0 1 .176-.198 3 3 0 1 1 3.847 0ZM9.2 7.073h-.001a1.206 1.206 0 0 0-2.398 0L6.305 11.5h3.39ZM9.5 3a1.5 1.5 0 1 0-3.001.001A1.5 1.5 0 0 0 9.5 3Z"],"h":16,"w":16}}
 export const accessibilityInset: OcticonSymbolVariants = {"16":{"p":["M8 0a8 8 0 1 1 0 16A8 8 0 0 1 8 0Zm2 4a2 2 0 1 0-2.95 1.76 1.87 1.87 0 0 0-.32.24H3.75a.75.75 0 0 0 0 1.5h2.363l-.607 5.67a.75.75 0 1 0 1.49.16l.25-2.33h1.508l.25 2.33a.75.75 0 0 0 1.492-.16L9.888 7.5h2.362a.75.75 0 0 0 0-1.5H9.27a1.98 1.98 0 0 0-.32-.24A2 2 0 0 0 10 4Z"],"h":16,"w":16}}
