+++ script/generate-octicons.ts
@@ -11,8 +11,6 @@ import * as cp from 'child_process'
 import { check } from 'reserved-words'
 import toCamelCase from 'to-camel-case'
 
-import { OcticonSymbolVariants } from '../app/src/ui/octicons'
-
 // Basic type for the @primer/octicons package.
 type OcticonsLib = Record<
   string,
@@ -37,12 +35,18 @@ type OcticonsLib = Record<
   }
 >
 
+interface IOcticonVariant {
+  p: string[]
+  h: number
+  w: number
+}
+
 interface IOcticon {
   /** JS friendly name of the icon. */
   readonly name: string
 
   /** Size variants of the icon. */
-  readonly variants: OcticonSymbolVariants
+  readonly variants: Record<PropertyKey, IOcticonVariant>
 }
 
 function getJsFriendlyName(name: string) {
@@ -62,7 +66,7 @@ async function generateIconData(): Promise<ReadonlyArray<IOcticon>> {
         throw new Error(`Unexpected empty sizes array for ${octicon.symbol}`)
       }
 
-      const variants: OcticonSymbolVariants = {}
+      const variants: Record<PropertyKey, IOcticonVariant> = {}
 
       Object.entries(octicon.heights).forEach(([height, data]) => {
         variants[height] = {
@@ -86,13 +90,25 @@ generateIconData().then(results => {
     }
   )
 
-  out.write(
-    '/*\n' +
-      ' * This file is automatically generated by the generate-octicons tool.\n' +
-      ' * Manually changing this file will only lead to sadness.\n' +
-      ' */\n\n' +
-      "import { OcticonSymbolVariants } from '.'\n\n"
-  )
+  out.write(`/*
+* This file is automatically generated by the generate-octicons tool.
+* Manually changing this file will only lead to sadness.
+*/
+
+export type OcticonSymbolVariant = {
+  /** SVG path element data */
+  readonly p: string[]
+
+  /** The width of the symbol */
+  readonly w: number
+
+  /** The height of the symbol */
+  readonly h: number
+}
+
+export type OcticonSymbolVariants = Record<PropertyKey, OcticonSymbolVariant>
+
+export type OcticonSymbol = OcticonSymbolVariant | OcticonSymbolVariants\n\n`)
 
   results.forEach(icon => {
     out.write(
